Files List drag-and-drop upload
- Implement drag-and-drop file upload for entire Project Files List content area
- Change "Upload Files" button in header of files list table to open the system
  UI file-selection interface instead of a modal form
- Move project files uploading logic to custom hook so it can be shared between
  components
