+++ src/scenes/Projects/Files/FileCreateActions.tsx
@@ -1,25 +1,32 @@
-import { Box, IconButton } from '@material-ui/core';
+import { IconButton } from '@material-ui/core';
 import { CreateNewFolder, Publish } from '@material-ui/icons';
 import React from 'react';
+import { useDropzone } from 'react-dropzone';
 import { useDialog } from '../../../components/Dialog';
 import { CreateProjectDirectory } from './CreateProjectDirectory';
-import { UploadProjectFiles } from './UploadProjectFiles';
+import { useUploadProjectFiles } from './useUploadProjectFiles';
 
 export const FileCreateActions = () => {
   const [createDirectoryState, createDirectory] = useDialog();
-  const [uploadFileState, uploadFile] = useDialog();
+  const handleFilesDrop = useUploadProjectFiles();
+
+  const { getRootProps, getInputProps, open: openFileBrowser } = useDropzone({
+    onDrop: handleFilesDrop,
+    noClick: true,
+  });
+
   return (
     <>
-      <Box>
-        <IconButton onClick={uploadFile}>
+      <div {...getRootProps({ className: 'dropzone' })}>
+        <input {...getInputProps()} />
+        <IconButton onClick={openFileBrowser}>
           <Publish />
         </IconButton>
         <IconButton onClick={createDirectory}>
           <CreateNewFolder />
         </IconButton>
-      </Box>
+      </div>
       <CreateProjectDirectory {...createDirectoryState} />
-      <UploadProjectFiles {...uploadFileState} />
     </>
   );
 };
